move file and copy-nit
